Release version 0.0.3

+++ src/deepphenotree/models.py
@@ -22,17 +22,16 @@ if __name__ == "__main__":
     out_dir.mkdir(parents=True, exist_ok=True)
 
     for share_url,folder in zip(urls,folders):
-        
         download_url = share_url.rstrip("/") + "/download"
 
         out_dir = Path(os.path.join(homedir,path,folder.replace('.pt','')))
         out_dir.mkdir(parents=True, exist_ok=True)
 
         out_path = homedir / out_dir / "best.pt"
-
-        with requests.get(download_url, stream=True, timeout=60) as r:
-            r.raise_for_status()
-            with open(out_path, "wb") as f:
-                for chunk in r.iter_content(chunk_size=1024 * 1024):
-                    if chunk:
-                        f.write(chunk)
+        if not out_path.exists():
+            with requests.get(download_url, stream=True, timeout=60) as r:
+                r.raise_for_status()
+                with open(out_path, "wb") as f:
+                    for chunk in r.iter_content(chunk_size=1024 * 1024):
+                        if chunk:
+                            f.write(chunk)
